pipederm/source/requirements.txt
# Core Framework
Flask==3.1.1
flask-cors==6.0.0
flask-limiter==3.5.0
Werkzeug==3.1.3

# API Documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# Deep Learning & AI
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.53.3
accelerate==1.6.0

# ADAPTADO: Dependências para o modelo melaNet (Keras/TensorFlow)
tensorflow>=2.10.0
keras>=2.10.0
h5py>=3.8.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless==4.12.0.88
numpy>=1.24.0

# Ollama Integration
ollama==0.5.1

# Background Tasks
APScheduler==3.11.0

# Development & Testing
pytest==8.4.1
python-dotenv==1.1.0

# Production Server
gunicorn==23.0.0